python/flask-chatgpt-getting-started/requirements.txt
Flask
openai
python-dotenv